+++ Source/SelfService/Web/micoservice/businessMomentsAdaptor/configuration/BearerAuthComponent.tsx
@@ -5,7 +5,7 @@ import { TextField } from '@fluentui/react/lib/TextField';
 import { Stack } from '@fluentui/react/lib/Stack';
 import { Label } from '@fluentui/react/lib/Label';
 
-import { ConnectorWebhookConfigBearer } from '../store';
+import { ConnectorWebhookConfigBearer } from '../../../store';
 
 const stackTokens = { childrenGap: 15 };
